[IMP] web_gantt: 'form_view_id' attribute
Gantt view supports forcing the form view to use when opening a record (on
pills click). To do so, we needed to add the context key `form_view_id` on
the action. This can cause problem when the context needs to be evaluated
python side (convert xmlid to id for the view), and webclient side (date
filter, for instance). Moveover, as the context can be propagated, this
context key might cause problems.
This commit transform this feature by adding an explicit attribute on gantt
view to support this use case, in a safer way.

Task-2050019

closes odoo/enterprise#5080

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>