[IMP] l10n_cl_reports: add F29 proposal
As the government receives all information already,
the F29 'tax return' does not need to be sent to the
government, but is a handy way to check things.

We need however to send this report a parameter.
(has to do with amounts used for private and client usage)
That is why there is a wizard.

closes odoo/enterprise#5257

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>