[IMP] l10n_cl_reports: add F29 proposal report
As the government receives all information already,
the F29 'tax return' does not need to be sent to the
government, but is a handy way to check things.

We need however to send this report a parameter.
(has to do with amounts used for private and client usage)
That is why there is a wizard.

[FIX] re-add coding utf-8

closes odoo/enterprise#5820

Signed-off-by: Josse Colpaert <jco@openerp.com>