[IMP] industry_fsm: make working with products easier
There were some issues with the way products are handled in fsm, regarding access rights and ways to create/edit products from within the fsm app.
This commit introduces the following changes to make it easier/better:

*Use price lists to compute products price

After this commit, the price shown on the product kanban and on the tasks tat button is computed based on product price + customer's price list. However given the way the sale order is generated + updated when product quantities changes, once a product was registered it's unit price and the total for that product won't be updated.

This is better as it avoids modifying a sale order which might already be in non draft stage.

*Allow all users to see products and worksheets linked to an fsm task

Prior to this commit, when having a sale access right lower or equal to sales: own document only, you would run into an access error when accessing a task on which products where linked by someone else. This is due to the fact when the first product is linked to a task, a sale order is created, of which the owner is the current user.

Then when another wants to open that task, some fields were to be computed in order to display a product stat button (which depend on related sale order). But since the actual user was not the owner, he would encounter an access error. In order to avoid this, the computations are done with sudo (the user can see what the number of products and at which price were linked).

The same problem occurs when accessing a worksheet, since it is also linked to the sale order. Again we added a sudo in a compute field.

Also for the user to do a full fsm flow, he needs to be able to create and edit a sale order, so we added sudo in the _fsm_create_sale_order method. But if the user didn't have the rights to do it himself, we don't set him as the salesman. So that the sale order don't get lost being assigned to someone who is unable to validate it.

*Allow to create products on the product kanban from fsm

When adding material to a task, if one want's to link a new product to
the task. He then has to go back to project, then
field service, add it's product and then go back to the task.

This commit adds a link to the product form on the material form,
when product is created, will be linked to the task's project.

The new field on product template was created to use in context with
default_ when going on the product form, so that it will be linked to
the project on save.

*Add a smiling face placeholder on material form

When installing the fsm app, the field service project is initially
empty. So when the user goes to a task to add product he lands
on an empty kanban. This commit adds the smiling face placeholder to
guide him to the next step.

related task 2000719