[ADD] helpdesk_fsm: allow to convert helpdesk ticket into fsm tasks
This module adds a settings to helpdesk teams that allows them to create fsm tasks based on a helpdesk ticket.
This is done through a new settings in helpdesk teams forms. There is also a stat button on the helpdesk ticket form that states how many and redirects to a list of the fsm tasks that were created through helpdesk.

On the other side there is also a stat button on fsm tasks to see the original ticket.

There is at most one ticket for a task but multiple tasks may be generated from the same ticket.

related task 2000719