[IMP/REF] sale_subscription: extendable renewal flow
The goal of this commit is to make the control flow of subscriptions
renewal more flexible.

By introducing a computed field, the behavior can be extended in other
modules, if need be.

This is useful for acquirers whose transaction's acceptance criteria is
somewhat different than 'done' or 'authorized' states.

task-1870363
closes odoo/enterprise#2531