[FIX] l10n_mx_edi: unit tests
- _onchange_product_id() was changed because when is executed change the
price line and not update the line_ids on the invoice, the returns the
error: UserError: ('Cannot create unbalanced journal entry. Ids: [56]')

- test_account_tax_cash_basis is improved to consider that the invoice
is the account.move and avoid write that value on the invoice.
Extra: lint error was fixed.

- l10n_mx_edi_external_trade, l10n_mx_edi_cancellation and l10n_mx_edi_payment_bank
modules not exists, then, the CFDI expected must be search in l10n_mx_edi.

- The fiscal position now is a field in the company, and is added with
data demo, then, is not necessary assign that value in the tests.

- action_invoice_open() method now not exists in the invoice, must be
used the post() method.

closes odoo/enterprise#5325

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>