[ADD] l10n_mx_edi: Add crons to update SAT status of invoices & payments
This adds crons to update daily the SAT status of invoices and payments
when their status is not all set.

When synchronizing the SAT status of a lot of records, e.g. all pending
invoices, it could lead to a systematic failure due to CPU/real-time
exceeded, thus being unable to complete the task.

To solve the above case, this also introduces two new methods to update
the SAT status by batches, one for invoices and another one for
payments; ensuring changes are committed after each batch is processed.
This is related to the following comment:
https://github.com/odoo/enterprise/pull/2207#issuecomment-392673989

closes odoo/enterprise#5299

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>