[IMP] planning: overlap detection
This commit provides an overlapping detection of planning slots. Purpose
is to ease people building the planning: they can now see which shifts are
overlapping the current one, in order to correct the agenda.
This is done with SQL query directly, which offers special tools to do it
in a faster way than using Odoo `search`.

Task-2024216