[IMP] planning: templating feature
The commit brings the "templating" feature. Purpose of this is to ease manager
to create the planning for their team; when saving a shift, their can register
it as a template. A template is a start hours for a role (optional) with a
duration.
The last used template are displayed above the form view when creating a shift.
Choosing one will prefill the start and end time and the role of the shift to
create.
A menu is dedicated to allow manager to edit, create or remove shift template.

Technically, we use a dummy field to trigger the onchange when selecting the
template. We don't need that informations so we don't keep it in database.

Task-2024216