IMP] pos_restaurant: multi-user sessions
PURPOSE
=========

When working in restaurants, there are often more than one waiter working on the same floor.
But Odoo doesn't support that.
There can only be one person working on one session for the same tables.
The purpose of this task to allow several people to work on the same session (while synchronizing sessions).

TASK-ID: 1891130

closes odoo/enterprise#5174

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>