[FIX] web_gantt: send datetime in the expected format
The string received by `gantt_unavailability` was not in a valid
datetime format for the server, this would cause a crash if the code was
only using `fields.Datetime.from_string` in order to convert the
received string.

To avoid this crash https://github.com/odoo/enterprise/commit/52228b9cfa16126e059c53fe436d3745fc71bdf1#diff-ea773c599012668663043126fb43b754R238 introduced a work arround
(using `.replace('T', ' ')`).

This work arround cannot be considered valid as it is the client
responsibility to send a valid information to the server (in our case a
valid datetime string).

This fix reuses the `convertToServerTime` method defined in
`gantt_model.js` in order to convert the datetime in the right format when
asking for `gantt_unavailability`.

It should be noted that there are already a number of functions in order
to convert from and to dates, times, datetimes and moment.js objects,
those are defined in `time.js` and should probably be considered as
replacements for the custom `convertToServerTime` method.

closes odoo/enterprise#5198

Signed-off-by: Romain Libert (rli) <rli@odoo.com>