[FIX] web_mobile: user image no more visible on mobile native app
Due to the changes introduced in odoo/odoo@58a2ffa26f1a3b0f9630ce16d11b758d18e20a21
the image_medium field was renamed into image_128.

But in the mobile native app we get the image of the user using this URL
"/web/image?model=res.users&field=image_medium&id=". We can't change this
URL because there is one mobile app for all versions of Odoo.

For now our app doesn't know which version of Odoo is used.
We want to avoid to increase code complexity.
It's why we preferred to add this backward compatibility fix.

closes odoo/enterprise#5187

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>