[FIX] hr_payroll_account: no company on payroll
When no company is set on the payslip, journal entries of zero are
generated.

This is because `currency` is empty, therefore `currency.round()` rounds
all amounts to zero.

We fall back on the company of the journal, since the latter is
mandatory.

opw-2054781

closes odoo/odoo#35859

X-port/Cherry-pick of odoo/odoo@bdf1817d0340c9ef8f2bdb0f95309c61c00b2208

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>