[IMP][REF] project_forecast: make depends on planning
Forecast and planning seems the same feature. This commit makes the forecast
application using the planning features and models. We consider planning as
small shift (usually the same day) and forecast (for longer period).

Some feature are not relevant for planning (the effective timesheeted
time, ...), so those feature are hidden. All labels are renamed into
'planning' to make the user understanding better.

With this commit, shift can be linked to project and task, in order to
handle the 'old' forecat features, but also to have a more advanced
planning feature by sorting shifts with another granularity level (project
or/and task).
In project app, the gantt view is different from the one in 'planning'
as we need to support the consolidation with percentage notation.

Task-2024216