[FIX] iot: Progress bar scan range
When opening the modal to scan the network, we added listeners to the
window to receive events during the scanning process. These listeners
were not removed when the modal was closed so reopening the modal used
to create a second listener with the same function. All listeners were
then executed twice, and the progress bar went up to 200%.

We need to remove the listeners when the modal is closed. When calling
removeEventListener, the reference of the function passed needs to be
the same that was added in the first place. Since we use `.bind(this)`,
we need to save this reference, otherwise it would change.

TaskID: 2056922

closes odoo/enterprise#5216

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>