[FIX] iot: Setup IoT Box
The setup of IoT Boxes failed since the merge of the task about
barcode scanners: odoo/odoo#35084

We tried to get the list of available keyboard layouts from the IoT
Box. The request was sent from the server, which cannot reach the Box
because it's not located on the same server.

We now perform this task the other way around. When a device of type
printer is created on the Box, we send the list of available layouts
to the server.

TaskID: 1961025

closes odoo/enterprise#5224

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>