[REF] account_reports[_followup]: Clean event bindings, template redefined
1. Event Bindings

Currently, buttons event handlers in the followup
view are attached using jQuery.
Instead, the Odoo style event binding should be
used instead. The syntax is shorter and declarative, hence
less error prone.

2. Template Redefined

The template `CustomerStatements.buttons` from `account_reports`
was not used in `account_reports_followup` and completely
redefined to add a button (template `CustomerStatements.buttonsWithLevels`).
Extending the first template is the correct way to add a button.
It is shorter, does not require to re-attach event handlers and
allows to remove the `renderButtons` override.

3. Button class

In `account_reports_followup`, buttons like "Send By Email",
"Print Letter", etc are set to primary if the current followup line
uses email, letter, etc, and secondary otherwise.
The logic to add or remove the classes was repeated several times
Now, it is factorised with a helper function.
This helper function will be reused in the future to send SMS
text messages.