[ADD] account_consolidation: new app to make financial consolidation
This module allows you create a consolidation chart of accounts, define the mapping with accounts from different companies and make the consolidation for these accounts in a single report. You can tweak the different currency rates to consider or the consolidation rate for each company.

It brings two new enterprise modules :
- account_consolidation with the business code
- l10n_be_us_consolidation_demo with demo data

Was task #1857935

closes odoo/enterprise#5773

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>