[IMP] account_reports: Add SMS support
Purpose
=======

In the follow-up reports, allow to send an SMS text
message to remind a partner he has some due payment.

Specification
=============

Add a primary button "Send By SMS" in the follow-up report form view to send
and SMS. The button opens the sms composer to send the message.

The message should be pre-filled with a default value defined on
the user's company.

Once the SMS has been sent, the buttons becomes secondary
and the button "Done" appears.

The SMS should appear in the chatter.