[IMP] web_grid, project_forecast: allow to use day as range in grid
Before this commit, the grid view only had week and month range.
Now grid_view also have day range.

"timesheet_grid" displayed week as default because it is the first one.

On mobile, we want to display it per day by default even if it's not
specified in the arch view.
Note that in desktop mode the day button has not been added.

We also had to add a dropdown to display grid ranges on mobile.

Task ID : 32285

Co-authored-by: Shubham Johar <sjo@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>