[FIX] web_studio, *: hide technical field widgets in non-debug mode
*:
- marketing_automation
- mrp_workorder
- quality
- timesheet_grid
- web_enterprise/static/src/js/views/upgrade_fields.js

This commit removes description of fields that should not be shown in
non-debug mode Studio. Only field widgets whose own description is
set are displayed in non-debug mode Studio in "Widget" of sidebar.
In other words, even if any ancestor has a description but not the
current widget, it won't be displayed in the sidebar in non-debug
mode.

Task-Id 2056992

closes odoo/enterprise#5213

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>