[REF] planning: percentage and allocation time
Before this commit, people could create forecast by setting the number of
allocated hour during the forecast. This computes the percentage of working
time of the employee. Now we inverse this: we fill the percentage and it
computes the number of allocated hours.
Now that we can do both planning and forecast, the computations are
different:
- if the shift is less then 24h (planning): the allocated hour is computed as
shift duration.
- if the shift last multiple days, the allocated hours is computed with the
working calendar of the employee.

The percentage also change the allocated hours. By default, it is 100%.
As we consider the percentage as an advance feature, the percentage field is
hidden with a technical group. This group is automatically activated when
installing project forecast to enjoy full forecast feature.

Task-2024216