[FIX] sale_amazon: remove the forced generation of invoices
Prior to this commit, an invoice was forcibly generated as soon as an order was synchronized from Amazon. This had for consequence to split the invoicing of Fulfilled by Merchant orders in two invoices if the order contained a service product and another product invoicable on delivery.

This commit also hides the stat button of Amazon Offers on products' forms if there is no offer for these product.

In addition, the term 'Enter your Amazon account' is replaced by 'Register your Amazon Account' in helper texts to emphasize the fact that accounts are created on Amazon Seller Central before being integrated in Odoo.

closes odoo/enterprise#5242

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>