[FIX] web_map: fix module name in res_config_settings
When a settings view is opened with the web_map module installed, the
client tries to load the description icon from
/map/static/description/icon.png , causing a 404 which can last 2 sec.

This makes the click_all test crash randomly when the total loading of
the page takes more than 5 sec.

With this commit, the correct web_map name is used in the view.

closes odoo/enterprise#5248

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>