[MOV] web_{mobile,enterprise}: move responsive code to web_enterprise
This commit moves responsive SCSS code which doesn't depend on the
mobile application (Android or iOS) out of the `web_mobile` module.

The purpose is to separate the code related to the mobile application
(cf. native features) and the code that makes the application more
responsive (cf. "on a small screen").

This is part of an ongoing effort to make this separation clearer :

* `web` responsive => `web_enterprise`
* `web` native => `web_mobile`