[FIX] pos_iot: do_action on widget
The widgets that send balance to payment terminals, get actions as
response from the server when they validate the pos session.

The actions returned were not handled, and just a reload was done, which
leads to not validate the session when the wizard was returned.

So we've converted the widget in an Abstract action and call do_action
on the returned action from the server.

TASK-ID: 2057436

closes odoo/enterprise#5251

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>