[IMP] industry_fsm: validate so on mark as done
This ensure the sale order invoice status stays consistent with fsm's flows. It is better than considering task in fsm_state==done and invoice_count > 1, because this covers the case of a downpayment or multiple invoices. Otherwise not fully invoiced tasks wouldn't be see in the to invoice filter.

Task-2047389      Closes #5388