[IMP] industry_fsm: make fsm default project depends on it's sequence
We now use the sequence field of fsm projects to determine which one should be proposed on the task form.
This do not change much in the default use case (one project for all tasks). Since it is the only one, it will be selected regardless of it's sequence.
In the use case where you use multiple projects, it is useful to be able to set a default project for fsm.

Task-2047389      Closes #5388