[FIX] web_studio: non-text-muted sidebar options dropdown items
Before this commit, dropdown options in studio sidebar had grey
color and did not look very well because of the white background.

This commit fixes the issue by making the color dark for sidebar
dropdown option and thus making the design more obvious.

Note that styling on the options are OS and browser specific. For
example, styling of <option> tags is not currently supported by
Safari, so it will always have the default appearance.

Task-Id 2057762

closes odoo/enterprise#5372

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>