[FIX] stock_barcode: open barcode for empty inventory
Before this commit, you can't enter in the barcode view of an inventory
if this inventory has no locations and no inventory lines.
How to reproduce:
    - Enable multilocation;
    - Create a new inventory adjustement with no locations and for one
      or more products with no quants;
    - Start the inventory;
    - Click on barcode stat button.

To fix that, we start the inventory from the location warehouse in this
case, and the user can scan a location's barcode to change the
inventoried location.

Task #2059465