[IMP] web_studio: preset the appropriate field in report editor
When adding a field in report or defining the visibility on any node with help
of domain, 95% of the time, user will be using field on the model itself.

For better usability, we now pre-select the field that is representing the
current document (or in case of adding fields in structures with loops like
data table, the looping variable itself), so that user does not have to
choose among quite technical items.

task - 2058702

closes odoo/enterprise#5370

Related: odoo/odoo#39548
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>