[FIX] l10n_mx_edi: Field l10n_mx_edi_cfdi_certificate_id was removed
In the next case:
1. Configure the company certificates
2. Create and sign a customer invoice

After this, the certificates for the company expire, and the SAT provide
a new certificates.

3. Try cancel the invoice in step 2, that was signed with the expired
certificates.

In this case, the CFDI must be cancelled with the new records, then is
not necessary save the Certificate used to sign an CFDI because when try
to cancel is necessary get the certificate valid.

closes odoo/enterprise#5326

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>