[FIX] documents: ensure tags name are up-to-date
Go to Documents / Configuration / Workspaces, choose a workspace
to edit (or create a new one), edit (or add) a facet in the one2many
list, in the dialog, edit (or add) a tag, click on 'Save & Close':
the tag names aren't updated in the main form view.

The reason is that they are displayed as many2many_tags, so their
'display_name' is displayed, whereas in the list in the dialog, we
are editing their 'name'.

This rev. adds the 'display_name' in the dialog (invisible), so
that the onchange is triggered and the webclient is aware of the
new 'display_name' to display in the many2many_tags.

Task 2060337

closes odoo/enterprise#5344

Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>