[IMP] web_gantt: ignore date and datetime group by on gantt
The gantt view is not able to use aggregated fields right now. At different places in the code, the name of the field currently grouped by is used to access data inside of objects.
Even if this could be changed and the view would work, it doesn't really make sense to group by date/datetime inside a view that already displays record according to date (and it would be very hard to read).
Therefore the group by date is currently ignored.

Task-2047389      Closes #5388