[FIX] web_enterprise: enable scroll on form view in studio
Before this commit, long form view were not scrollable in studio.
As a result, it was not possible to properly edit the view.

This issue was introduced with https://github.com/odoo/enterprise/commit/97ebb077fc5cc7f81927d519919b84f4c4030fc7

This commit fixes the issue by removing the `overflow: hidden;`,
which does not seem useful.

Task-Id 2057746

closes odoo/enterprise#5415

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>