[FIX] industry_fsm: fix stat button showing products that were removed
When adding products and timesheet on an fsm task, these are added on it's related sale order.
When an fsm task is mark as done, it's sale order is confirmed. But in fsm one can still add/remove products and timesheets on a 'done' tasks. We then modify the confirmed so.
This yields problems when removing products (when their quantities reach 0), as the sale order is confirmed we cannot unlink the SOL, we write 0 on the quantity. Since the stat button on the task form counted the number of line, the product that had been removed this way were still counted in, this commit fixes it by filtering lines that have a positive quantity.

Task-2047389      Closes #5388