[IMP] industry_fsm: fix access rights issues on invoice generation
When reaching the end of the flow for an fsm task, there is the possibility to invoice time and material linked to this task. In order for the invoice to be generated the user needs the billing group, but in order for the material to be on the invoice, the user also needs to be a sales user with access to all documents (since the generated SO behind the task is owned by the first user to add time and/or products to the task, so own documents might not be sufficient).

Since we have no easy/good way to hide the create invoice button and to invoice menu, if you don't have both groups (when mixed, they are mixed with a logical or, we need a logical and), the new strategy is to have the fsm manager group imply sale and billing, and show these buttons/menus only to the fsm manager. This group is required anyway to have access to other menus than My Tasks inside fsm.

Task-2047389      Closes #5388