[FIX] helpdesk: Use correct field to compute exceeded_days
Currently, a Public user cannot submit a ticket.
He is faced with an access error.

The reason being the computation of field `exceeded_days` of
the `helpdesk.sla.status`.
In the compute method, a value was never assigned to the field
(wrong field name was used).
This later leads to a CacheMiss which triggers a read in database
and an access rights check.
A public user doesn't have read access rights to `helpdesk.sla.status`,
hence the crash.

Note: in a normal flow, the compute method is `compute_sudo`
which explains why it doesn't raise an AccessError for a public user.