[FIX] web_studio: improve chatter preview design
Currently, chatter preview design inside the chatter hook on form
view is not consistent with the actual chatter. So, few of the css
rules for chatter are not applied properly and the UI is ruined. The
issue did not come into light due to float property defined on
followers which was removed in commit
https://github.com/odoo/odoo/commit/ece2fed1545faa388777227d3a7525cf99ab1217#diff-920b56b15af5bcb7ab87e7cb296da98eR1.

This commit fixes the issue making the chatter preview design more
consistent with the actual chatter and thus reducing the
possibilities of breaking UI of the chatter hook in studio.

task-2060874

closes odoo/enterprise#5813

X-original-commit: 7e3610faaaf5cd7d9b7282c0b43fd5851a455865
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>