[FIX] web_enterprise: mobile: scrolling issue after input focus
This issue only happens with Chrome Android and Bootstrap 3-4 modals.
It was fixed in chrome 66 but it seems that it has been reintroduced
between Chrome 72.0.3626.121 and 73.0.3683.90.
So, it's still present as of today (Chrome Canary 78.0.3901.0).

Steps to reproduce:
- Go to Sales
- Open a quotation
- Send message
- Open it in full screen mode
- Click on Subject field
=> The view scroll to the wrong place and you will not be able to
see the field you are editing.

To deal with this, we found a workaround by adding "position: fixed" on
the document body when the modal is opened. In this way, the scroll will
be blocked under the modal. To achieve this Bootstrap provides a "modal-open"
class on the document body as we can see in :
https://github.com/twbs/bootstrap/commit/01405523229f3b9750326aa832716ad12009cf4b

We also noticed that we were not impacted by this Chrome issue before the
refactoring of the control panel :
https://github.com/odoo/odoo/commit/40dd121938552d3f88289d2e5fcd1d57a0dd01bb
To be honest we don't know exactly which change in the DOM highlighted
this issue.

We already noticed Chromium team about it.

Links:
- Bootstrap Bug tracker :
- https://github.com/twbs/bootstrap/issues/17338
- Chromium Bug tracker :
- https://bugs.chromium.org/p/chromium/issues/detail?id=595210
- Standalone bug (with Bootstrap 4.3.1 and Chrome 76.0.3809.132) :
- https://output.jsbin.com/layiqom/quiet

closes odoo/enterprise#5400

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>