[IMP] planning: read_group expand
To ease encoding planning for manager, we want to display employee on gantt
view that have been assigned in the last 30 days (arbitrary value). As the
"read group expand" necanism is built in the ORM, we use it even if it only
works for one level of groupby.
This expand is activated with a context key to avoid to overload other views
or actions.

Task-2024216