[ADD] social_sale: Add quotations/revenues statistics in social
This commit adds the quotations/revenues statistics to the social.post.

Whenever a quotation or an account.move gets created thanks to a tracked link in a social post,
its utms will be set this way:
 - utm campaign will be the campaign linked to the post (if any)
 - utm medium will be medium linked to the account to which the post was posted to.
 - utm source which will be the source linked to the social post itself.

In a way, we could say that the medium is the account while the source is the social.post itself.

This allows us to easily compute the number of quotations or the amount of revenues
generated by a particular post or even by account if it ends up being a required feature.

LINKS:

TaskID: 2045210
PR: #5406