[REF] planning: simplify recurrence
This commit brings a refactoring of the recurrency implemented previously
in forecast. We keep a dedicated model to handle the recurrence. Its
reponsibility is to generate shifts and remove shifts.
We simplify the notation by only keeping recurring shifts by week and until a
certain date, or forever. A shift can be set as recurrent directly on its
creation or modification. When a shift is not recurrent anymore (unchecking
'repeat' flag), the following shifts are remove, such as the recurrence itself.
The first shift of the recurrence is used as model to copy its values and
generate its siblings. Empty recurrences are removed with the cron.

The decision was made to remove the wizard creating the recurrence and the shift,
but to allow the shifts form view showing the recurency fields, we used computed
field with inverse method creating the recurrence. As those fields are there only
for usability and UI purpose, it is normal not to store them.
Tests were adapted and wizard removed. Code is now simpler and smaller.

The cron frequency is also simplified by hardcoding the unit (month). The
interval is still customizable by company, but is hidden in debug mode in
planning settings, as this is technical.

Task-2024216