[FIX] web_studio: prevent flickering on list views
Since commit [1], the css rules that were defined to prevent
flickering were broken and there was flickering on the list view
while hovering in the studio.

This commit fixes the issue by reverting the unnecessary changes
introduced in above mentioned commit and thus avoiding the flickering
on list view.

[1] https://github.com/odoo/enterprise/commit/3868e02a15af95552f98da80dc4640978c5b8854

task-2063189

closes odoo/enterprise#5478

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>