[IMP] project_timesheet_synchro: display remaining hours on task in dropdown
When timesheeting, it is useful to know how many hours are left considering already timesheeted time and forecasted hours. In other places in odoo this is done through the nameget of a task + a context key to avoid polluting every views. But since the timesheet app relies on import/export to fetch/upload data, the name isn't adapted and forcing it to do so wouldn't be a great idea as these dynamic names might break the import mechanism.

Here we export the extra field remaining_hours when fetching tasks and simulate the same mechanism as the nameget but in js code. There was however an issue with synchronisation, as the behaviour is to first fetch existing tasks and project from the backend then uploading timesheets that were created through the app. So the new timesheets are not accounted for when the remaining_hours field is computed. It is showing the state of the backend before syncing.

In order to work around this without changing the whole synchronisation code, we manually compute what the remaining_hours should be when fetching remote tasks. This is done by considering the amount of timesheet that are to be uploaded before syncing and removing that quantity from the task's remaining hours.

related task 1856139

closes odoo/enterprise#4400

Related: odoo/odoo#33586
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>