[IMP] pos_iot: Remove payment terminal from POS config
A "Payment terminal" field was still present in POS Config. This
configurations has been moved to the payment lines.

When closing the session of the POS we still used this configuration
and sent a balance operation to a single payment terminal only.
We now look through all of the payment lines and perform the balance
on all used terminals.

closes odoo/enterprise#5108

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>