[IMP] pos_iot: Select devices instead of Box in POS config
Previously, to use IoT Devices in the POS, user had to select an IoT
Box and check the checkboxes corresponding to the devices they wanted
to use. The first device of the corresponding type that had been found
was used. We now offer the possibility to select precisely the device
you want to use.

This also adds the possibility to use the same device across multiple
POS.

TaskID: 1970884