[FIX] web_studio: no traceback on page reload
Before this commit, when reloading the page when studio was active,
it always crashed.

This commit fixes the issue by not rendering studio in case no
current studio action exists. This prevents the crash, and still
let studio being open automatically after a short delay (~ 1second),
similarly to version 12.

Task-Id 2058509

closes odoo/enterprise#5433

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>