[FIX] delivery: changing customer abroad raises error when updating
Issue: When a sale.order already has a delivery line for a certain local
customer, changing to foreign customer and updating the delivery line
results to an error.

Commit Description: This commit fixes the above issue by passing the
right delivery_nature to the check_required_value when rating the
delivery.