[FIX] hr_payroll_account: Use a company journal as default value
The default value for the `journal_id` of a payroll structure
is the first general account journal found. However,
if multiple companies are active, there is no garantee on the
journal company.
Payroll strucutes are country dependent, but this could lead
to a Belgian structure but with a journal from a US company.
This is obviously wrong.

For the record, the following fix was needed in order to
implement the default value with `company_dependant`: odoo/odoo#36545