[FIX] l10n_be_hr_payroll: Generate notice period before holiday attest
There are two departure wizard.
One to generate the notice period and the departure payslip
and the second for the employee's holidays.

The latter should be called after the first one.
Currently, when the holiday attest wizard is called first
a traceback pops up because the notice period is not set.