[IMP] pos_iot: Multiple IoT Boxes in status icon
Since we now select devices in PoS Coonfig, we could use devices from
multiple IoT Boxes in the PoS. The proxy status icon need to check the
status of devices on any Box and check the status of all Boxes.

The status of the devices is retrieved with a RPC. iot.device records
have a "status" field to know if they're connected. The value in this
field is valid if the Box is connected.

To check if a Box is connected, we check the result of every action or
listener that is added on this Box, if an error occurs, then the Box is
disconnected.

The user can also click on the proxy status icon to manually update the
status of the Boxes. This will ping all of the IoT Boxes that have a
device that is used in the PoS.

TaskID: 1970884