[FIX] delivery_fedex: return label is not printed
If fedex international delivery method is configured to generate return
label, validating the picking skips the printing of return label because
of the missing date in the response of process_shipment function. This
commit fixes this behaviour allowing the return label to be printed even
if there is no date in the response.

Additionally, this commit handles the error encountered when processing
return shipment.